DEV-3882: Abstract Portal logic to new component + fix custom styles portal
