--- spa/src/components/portal/PortalEntry/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PortalEntry';
